Split files. #3
