+++ client/index.js
@@ -1,37 +1,20 @@
 const Redux = require('redux')
 const React = require('react')
 const ReactDOM = require('react-dom')
+const NotesList = require('./NotesList')
+//const NoteForm = require('./NoteForm')
+const store = require('./store')
 const app = document.getElementById('container')
 
-let state = ['Babelify won\'t work without .babelrc file!',
-'let = values that can change, const = values that cannot be reassigned',
- 'falsy values = null, NaN, 0, empty string, false, undefined',
- 'fetch returns a Promise']
-
-function reducer(state, action) {
-  switch (action.type) {
-    case 'NOTE_ADDED':
-      return state.concat(action.text)
-    default: return state
-  }
-}
-
-const store = Redux.createStore(reducer, state)
-
-const NotesList = (props) => {
-  props = store.getState()
-  const notes = props.map((note) => {
-    return <li className="note"> {note} </li>
-  })
+/*function App(props) {
+  const { noteInput } = props
   return (
-    <div id="list-container">
-      <ul id="list">
-        <h2 id="title">My Notes:</h2>
-        { notes }
-      </ul>
+    <div>
+      <NotesList/>
+      <NoteForm/>
     </div>
   )
-}
+}*/
 
 function render() {
   let current = store.getState()
