+++ public/style.css
@@ -5,7 +5,7 @@ body {
 #list-container {
   position: absolute;
   top: 100px;
-  left: 35%;
+  left: 50%;
   width: 600px;
   border: 2px black solid;
   background-color: #ededed;
